requirements.txt
streamlit==1.35.0
google-generativeai==0.5.4
langchain==0.1.14
langchain-community==0.0.19
faiss-cpu==1.7.4
python-dotenv==1.0.1
pypdf==3.17.1
